+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add a get_local_ip function
+- Add BaseApp.has_option to check if an app has a given option
+- Add support for authenticated Steam installs
 
 ### Changed